Segregate class declare from definition + refactor
