--- ow_power_system/include/PowerSystemNode.h
@@ -1,10 +1,0 @@
-#ifndef __POWER_SYSTEM_NODE_H__
-#define __POWER_SYSTEM_NODE_H__
-
-#include <vector>
-#include <map>
-#include <string>
-
-#include "PrognoserFactory.h"
-
-#endif
